pyproject.toml
[tool.poetry]
name = "dogcity(django)"
version = "0.1.0"
description = ""
authors = ["Ilya Kibirev <kibirev.i@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "==3.6.0"
astroid = "==2.14.2"
attrs = "==22.2.0"
autoflake = "==1.4"
black = "==23.1.0"
cachecontrol = "==0.12.11"
celery-types = "==0.14.0"
certifi = "==2022.12.7"
cffi = "==1.15.1"
charset-normalizer = "==3.0.1"
cleo = "==2.0.1"
click = "==8.1.3"
crashtest = "==0.4.1"
cryptography = "==39.0.1"
dill = "==0.3.6"
distlib = "==0.3.6"
django = "==4.1.2"
django-admin-display = "==1.3.0"
django-cors-headers = "==3.13.0"
django-phonenumber-field = "==7.0.2"
django-stubs = "==1.15.0"
django-stubs-ext = "==0.7.0"
djangorestframework = "==3.14.0"
djangorestframework-stubs = "==1.9.0"
dulwich = "==0.20.50"
filelock = "==3.9.0"
html5lib = "==1.1"
idna = "==3.4"
importlib-metadata = "==6.0.0"
isort = "==5.10.1"
jaraco-classes = "==3.2.3"
jeepney = "==0.8.0"
jsonschema = "==4.17.3"
keyring = "==23.13.1"
lazy-object-proxy = "==1.9.0"
lockfile = "==0.12.2"
mccabe = "==0.7.0"
more-itertools = "==9.0.0"
msgpack = "==1.0.4"
mypy = "==1.0.1"
mypy-extensions = "==1.0.0"
packaging = "==23.0"
pathspec = "==0.11.0"
pexpect = "==4.8.0"
phonenumbers = "==8.13.6"
pika-stubs = "==0.1.3"
pkginfo = "==1.9.6"
platformdirs = "==2.6.2"
poetry = "==1.3.2"
poetry-core = "==1.4.0"
poetry-plugin-export = "==1.3.0"
psycopg2-binary = "==2.9.5"
ptyprocess = "==0.7.0"
pycparser = "==2.21"
pyflakes = "==3.0.1"
pylint = "==2.16.2"
pylint-django = "==2.5.3"
pylint-plugin-utils = "==0.7"
pyrsistent = "==0.19.3"
python-dotenv = "==0.21.1"
pytz = "==2022.7.1"
rapidfuzz = "==2.13.7"
requests = "==2.28.2"
requests-toolbelt = "==0.10.1"
ruff = "==0.0.251"
secretstorage = "==3.3.3"
shellingham = "==1.5.0.post1"
six = "==1.16.0"
sqlparse = "==0.4.3"
tomli = "==2.0.1"
tomlkit = "==0.11.6"
trove-classifiers = "==2023.2.20"
types-backports = "==0.1.3"
types-chardet = "==5.0.4.2"
types-docutils = "==0.19.1.6"
types-frozendict = "==2.0.9"
types-mock = "==5.0.0.5"
types-python-dateutil = "==2.8.19.8"
types-pytz = "==2022.7.1.2"
types-pyyaml = "==6.0.12.8"
types-requests = "==2.28.11.14"
types-setuptools = "==67.4.0.1"
types-ujson = "==5.7.0.1"
types-urllib3 = "==1.26.25.7"
typing-extensions = "==4.5.0"
urllib3 = "==1.26.14"
virtualenv = "==20.19.0"
webencodings = "==0.5.1"
wrapt = "==1.14.1"
zipp = "==3.14.0"
gunicorn = "^20.1.0"
pre-commit = "^3.1.0"


[tool.isort]
include_trailing_comma = true
line_length = 88
multi_line_output = 3

[tool.ruff]
ignore = ["E999", "E501", "F405", "F403"]

[tool.mypy]
check_untyped_defs = true 
disallow_any_unimported = true 
install_types = true 
no_implicit_optional = true 
no_implicit_reexport = true 
no_warn_no_return = true 
non_interactive = true 
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"] 
pretty = true 
show_error_codes = true 
strict_equality = true 
warn_redundant_casts = true 
warn_unreachable = true 
ignore_missing_imports = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = "devtools"

[[tool.mypy.overrides]]
ignore_errors = true
module = "*.migrations.*"

[tool.django-stubs]
django_settings_module = "config.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"